feat: implementing property use cases to get one, get all and update

+++ src/main/java/com/digisphere/propertize/application/property/useCase/ICreateProperty.java
@@ -6,5 +6,5 @@ import java.time.LocalDate;
 import java.util.Map;
 
 public interface ICreateProperty {
-    Property execute(String ownerId, String description, Map<String, String> address, String type, Double size, Integer bedRoomCount, Integer bathRoomCount, Boolean hasGarage, Integer rentValue, String status, LocalDate constructionDate);
+    Property execute(String ownerId, String description, Map<String, String> address, String type, Double size, Integer bedRoomCount, Integer bathRoomCount, Boolean hasGarage, Double rentValue, String status, LocalDate constructionDate);
 }
